# Database/Properties/swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"api"},"paths":{"/api/Contracts":{"get":{"tags":["Contracts"],"operationId":"GetContracts","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/ContractModel"}}}}}},"/api/Contracts/declined":{"get":{"tags":["Contracts"],"operationId":"GetDeclinedContracts","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/ContractModel"}}}}}},"/api/Contracts/{userId}":{"get":{"tags":["Contracts"],"operationId":"GetContracts","consumes":[],"produces":[],"parameters":[{"name":"userId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Contracts"],"operationId":"SuggestContract","consumes":["multipart/form-data"],"produces":[],"parameters":[{"name":"userId","in":"path","required":true,"type":"string"},{"name":"file","in":"formData","required":false,"type":"file"}],"responses":{"200":{"description":"Success"}}}},"/api/Contracts/getFile/{contractId}":{"get":{"tags":["Contracts"],"operationId":"GetFile","consumes":[],"produces":[],"parameters":[{"name":"contractId","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}},"/api/Contracts/sign/{contractId}":{"put":{"tags":["Contracts"],"operationId":"SignContract","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"contractId","in":"path","required":true,"type":"integer","format":"int32"},{"name":"sign","in":"body","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success"}}}},"/api/Contracts/Decline/{id}":{"put":{"tags":["Contracts"],"operationId":"DeclineContract","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}},"/api/Contracts/{id}":{"delete":{"tags":["Contracts"],"operationId":"DeleteContract","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}},"/api/Users/register":{"post":{"tags":["Users"],"operationId":"RegisterAsync","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"userRegisterRequestModel","in":"body","required":false,"schema":{"$ref":"#/definitions/UserRegistrationModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/Users/login":{"post":{"tags":["Users"],"operationId":"Login","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"model","in":"body","required":false,"schema":{"$ref":"#/definitions/AuthModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/Users/logout":{"post":{"tags":["Users"],"operationId":"LogOff","consumes":[],"produces":[],"parameters":[],"responses":{"200":{"description":"Success"}}}},"/api/Values":{"get":{"tags":["Values"],"operationId":"Get","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"type":"string"}}}}},"post":{"tags":["Values"],"operationId":"Post","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"value","in":"body","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success"}}}},"/api/Values/{id}":{"get":{"tags":["Values"],"operationId":"Get","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"type":"string"}}}},"put":{"tags":["Values"],"operationId":"Put","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"value","in":"body","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["Values"],"operationId":"Delete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}}},"definitions":{"ContractModel":{"type":"object","properties":{"contractId":{"format":"int32","type":"integer"},"userId":{"type":"string"},"documentId":{"format":"int32","type":"integer"},"documentName":{"type":"string"},"date":{"format":"date-time","type":"string"},"declinedDate":{"format":"date-time","type":"string"},"sign":{"type":"string"},"isRefused":{"type":"boolean"}}},"UserRegistrationModel":{"required":["email","password"],"type":"object","properties":{"role":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"password":{"format":"password","type":"string"}}},"AuthModel":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"format":"password","type":"string"}}}}}